move conditionValidator to search utils + import fixes
