+++ src/components/discovery/DiscoveryQueryBuilder.js
@@ -13,36 +13,11 @@ import {
 } from "@/modules/explorer/actions";
 import { useDataTypes, useServices } from "@/modules/services/hooks";
 import { useAppDispatch, useAppSelector } from "@/store";
-import { nop } from "@/utils/misc";
-import { VARIANT_OPTIONAL_FIELDS } from "@/utils/search";
-import { objectWithoutProp } from "@/utils/misc";
+import { nop, objectWithoutProp } from "@/utils/misc";
+import { conditionValidator } from "@/utils/search";
 
 import DataTypeExplorationModal from "./DataTypeExplorationModal";
 import DiscoverySearchForm from "./DiscoverySearchForm";
-import { getSchemaTypeTransformer } from "./DiscoverySearchCondition";
-
-const conditionValidator = (rule, { field, fieldSchema, searchValue }) => {
-  if (field === undefined) {
-    return Promise.reject("A field must be specified for this search condition.");
-  }
-
-  const transformedSearchValue = getSchemaTypeTransformer(fieldSchema.type)[1](searchValue);
-  const isEnum = fieldSchema.hasOwnProperty("enum");
-  const isString = fieldSchema.type === "string";
-
-  // noinspection JSCheckFunctionSignatures
-  if (
-    !VARIANT_OPTIONAL_FIELDS.includes(field) &&
-    (transformedSearchValue === null ||
-      (!isEnum && !transformedSearchValue) ||
-      (!isEnum && isString && !transformedSearchValue.trim()) || // Forbid whitespace-only free-text searches
-      (isEnum && !fieldSchema.enum.includes(transformedSearchValue)))
-  ) {
-    return Promise.reject(`This field must have a value: ${field}`);
-  }
-
-  return Promise.resolve();
-};
 
 const DiscoveryQueryBuilder = ({ activeDataset, dataTypeForms, requiredDataTypes, onSubmit, searchLoading }) => {
   const dispatch = useAppDispatch();
@@ -169,7 +144,6 @@ const DiscoveryQueryBuilder = ({ activeDataset, dataTypeForms, requiredDataTypes
               setFormRef={setFormRef}
               onChange={handleChange}
               handleVariantHiddenFieldChange={handleVariantHiddenFieldChange}
-              conditionValidator={conditionValidator}
             />
           ),
         };
